@@ -1,23 +1,26 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
-        classpath 'com.novoda:bintray-release:0.4.0'
+        classpath 'com.android.tools.build:gradle:3.3.0-alpha07'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
+        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
     }
 }
 
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
 
 ext {
-    compileSdkVersion = 25
-    buildToolsVersion = "25.0.2"
-    supportLibVersion = "25.3.1"
+    compileSdkVersion = 28
+    appcompatVersion = "1.0.0-rc01"
+    cardviewVersion = "1.0.0-rc01"
 }
 
 task clean(type: Delete) {
